Got rid of the example app project
