+++ LeanTest.sln
@@ -23,8 +23,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Build", "Build", "{62FABEA8
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "State.Examples.MsTest", "Source\State.Examples.MsTest\State.Examples.MsTest.csproj", "{A7470C2D-CB2B-4D63-BEE2-835381A29F00}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExampleApp", "Source\ExampleApp\ExampleApp.csproj", "{6E0992FC-C481-49C0-8A03-D36FE95F67C4}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MSTest", "Source\MSTest\MSTest.csproj", "{FBEBB99E-BE1E-45AC-94C1-DFBBCEC69E05}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MSTest.Examples.MsTest", "Source\MSTest.Examples.MsTest\MSTest.Examples.MsTest.csproj", "{BC15DD45-522D-4086-B7D1-63806896D417}"
@@ -80,10 +78,6 @@ Global
 		{A7470C2D-CB2B-4D63-BEE2-835381A29F00}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A7470C2D-CB2B-4D63-BEE2-835381A29F00}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A7470C2D-CB2B-4D63-BEE2-835381A29F00}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6E0992FC-C481-49C0-8A03-D36FE95F67C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6E0992FC-C481-49C0-8A03-D36FE95F67C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6E0992FC-C481-49C0-8A03-D36FE95F67C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6E0992FC-C481-49C0-8A03-D36FE95F67C4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FBEBB99E-BE1E-45AC-94C1-DFBBCEC69E05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FBEBB99E-BE1E-45AC-94C1-DFBBCEC69E05}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FBEBB99E-BE1E-45AC-94C1-DFBBCEC69E05}.Release|Any CPU.ActiveCfg = Release|Any CPU
