--- Source/ExampleApp/Domain/MyData.cs
@@ -1,13 +1,0 @@
-﻿namespace ExampleApp.Domain
-{
-    public class MyData
-    {
-        public string Key { get; set; }
-        public int Age { get; set; }
-    }
-    public class MyOtherData
-    {
-        public string OtherKey { get; set; }
-        public int OtherAge { get; set; }
-    }
-}
