+++ Source/Mock.Examples.MsTest/Domain/MyData.cs
@@ -1,0 +1,8 @@
+﻿namespace Mock.Examples.MsTest.Domain
+{
+	public class MyData
+	{
+		public string Key { get; set; }
+		public int Age { get; set; }
+	}
+}
