+++ Source/Mock.Examples.MsTest/Mocks/MockMyExternalService.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using ExampleApp.Domain;
 using LeanTest.Mock;
 using Mock.Examples.MsTest.Application;
+using Mock.Examples.MsTest.Domain;
 
 namespace Mock.Examples.MsTest.Mocks
 {
