+++ Source/Mock.Examples.MsTest/TestMyApplicationService.cs
@@ -1,7 +1,7 @@
-﻿using ExampleApp.Domain;
-using LeanTest.Core.ExecutionHandling;
+﻿using LeanTest.Core.ExecutionHandling;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Mock.Examples.MsTest.Application;
+using Mock.Examples.MsTest.Domain;
 
 namespace Mock.Examples.MsTest
 {
