+++ Source/State.Examples.MsTest/Domain/MyOtherData.cs
@@ -1,0 +1,8 @@
+﻿namespace State.Examples.MsTest.Domain
+{
+	public class MyOtherData
+	{
+		public string OtherKey { get; set; }
+		public int OtherAge { get; set; }
+	}
+}
