+++ Source/State.Examples.MsTest/TestMyApplicationService.cs
@@ -1,14 +1,14 @@
-﻿using ExampleApp.Domain;
-using LeanTest.Core.ExecutionHandling;
+﻿using LeanTest.Core.ExecutionHandling;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using State.Examples.MsTest.Application;
+using State.Examples.MsTest.Domain;
 
 namespace State.Examples.MsTest
 {
     /// <summary>Note that in a real-world example we must choose an IoC container - here we have implemented our own for
     /// this example.
     /// This container must be initialized in an AssemblyInitializer class - refer to this class to see how it is
-    /// sone with our example IoC container.
+    /// done with our example IoC container.
     /// For the example to be more complete, there should be two projects, one for test one for what is being tested.</summary>
     [TestClass]
     public class TestMyApplicationService
